+++ src/org/broad/igv/sam/AlignmentBlock.java
@@ -1,19 +1,12 @@
 /*
- * Copyright (c) 2007-2011 by The Broad Institute of MIT and Harvard.  All Rights Reserved.
+ * Copyright (c) 2007-2012 The Broad Institute, Inc.
+ * SOFTWARE COPYRIGHT NOTICE
+ * This software and its documentation are the copyright of the Broad Institute, Inc. All rights are reserved.
+ *
+ * This software is supplied without any warranty or guaranteed support whatsoever. The Broad Institute is not responsible for its use, misuse, or functionality.
  *
  * This software is licensed under the terms of the GNU Lesser General Public License (LGPL),
  * Version 2.1 which is available at http://www.opensource.org/licenses/lgpl-2.1.php.
- *
- * THE SOFTWARE IS PROVIDED "AS IS." THE BROAD AND MIT MAKE NO REPRESENTATIONS OR
- * WARRANTES OF ANY KIND CONCERNING THE SOFTWARE, EXPRESS OR IMPLIED, INCLUDING,
- * WITHOUT LIMITATION, WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- * PURPOSE, NONINFRINGEMENT, OR THE ABSENCE OF LATENT OR OTHER DEFECTS, WHETHER
- * OR NOT DISCOVERABLE.  IN NO EVENT SHALL THE BROAD OR MIT, OR THEIR RESPECTIVE
- * TRUSTEES, DIRECTORS, OFFICERS, EMPLOYEES, AND AFFILIATES BE LIABLE FOR ANY DAMAGES
- * OF ANY KIND, INCLUDING, WITHOUT LIMITATION, INCIDENTAL OR CONSEQUENTIAL DAMAGES,
- * ECONOMIC DAMAGES OR INJURY TO PROPERTY AND LOST PROFITS, REGARDLESS OF WHETHER
- * THE BROAD OR MIT SHALL BE ADVISED, SHALL HAVE OTHER REASON TO KNOW, OR IN FACT
- * SHALL KNOW OF THE POSSIBILITY OF THE FOREGOING.
  */
 
 /*
@@ -29,7 +22,7 @@ public class AlignmentBlock {
     private int start;
     private byte[] bases;
     public byte[] qualities;
-    private byte[] counts;
+    private short[] counts;
 
     private boolean softClipped = false;
     private Alignment baseAlignment = null;
@@ -87,15 +80,15 @@ public class AlignmentBlock {
         return qualities;
     }
 
-    public byte[] getCounts() {
+    public short[] getCounts() {
         return counts;
     }
 
-    public byte getCount(int i) {
+    public short getCount(int i) {
         return counts[i];
     }
 
-    public void setCounts(byte[] counts) {
+    public void setCounts(short[] counts) {
         this.counts = counts;
     }
 
@@ -131,15 +124,15 @@ public class AlignmentBlock {
     public String toString() {
         StringBuffer sb = new StringBuffer();
         sb.append("[block ");
-        sb.append(isSoftClipped()?"softClipped ":" ");
+        sb.append(isSoftClipped() ? "softClipped " : " ");
         sb.append(getStart());
         sb.append("-");
         sb.append(getEnd());
         sb.append(" ");
-        for (int i=0;i<bases.length;i++) {
-            sb.append((char)bases[i]);
+        for (int i = 0; i < bases.length; i++) {
+            sb.append((char) bases[i]);
         }
-        sb.append("]")  ;
+        sb.append("]");
         return sb.toString();
     }
 }
